Consolidate sub-resources under their parent
`Buyer` and `Refund` are sub-resources of `Invoice`, thus it is pragmatic to implement and test them within their parent's context.
