--- src/Actions/ManageBuyers.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace Vrajroham\LaravelBitpay\Actions;
-
-use BitPaySDK\Model\Invoice\Buyer;
-
-trait ManageBuyers
-{
-    /**
-     * @return Buyer
-     */
-    public static function Buyer(): Buyer
-    {
-        return new Buyer();
-    }
-}
