--- tests/LaravelBitpayRefundTest.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace Vrajroham\LaravelBitpay\Tests;
-
-use BitPaySDK\Model\Invoice\Refund;
-use PHPUnit\Framework\TestCase;
-use Vrajroham\LaravelBitpay\LaravelBitpay;
-
-class LaravelBitpayRefundTest extends TestCase
-{
-    /** @test */
-    public function isInstanceOfRefund()
-    {
-        $this->assertEquals(LaravelBitpay::Refund() instanceof Refund, true);
-    }
-}
